Move old config structs/methods into config_legacy.go
